requirements.txt
django
django-bootstrap-daterangepicker
django-crispy-forms
django-import-export
django-multiselectfield
crispy-bootstrap4
pillow